# Add optimisations for E2E tests