--- a/frontend/fingertips-frontend/Dockerfile
+++ b/frontend/fingertips-frontend/Dockerfile
@@ -65,4 +65,4 @@ ENV PORT=3000
 # server.js is created by next build from the standalone output
 # https://nextjs.org/docs/pages/api-reference/next-config-js/output
 ENV HOSTNAME="0.0.0.0"
-CMD ["node", "server.js"]
+CMD ["node", "--optimize-for-size", "server.js"]