… frontend/fingertips-frontend/package.json