# Description

Jira ticket: [DHSCFT-540](https://bjss-enterprise.atlassian.net/browse/DHSCFT-540)

Playwright e2e tests currently failing in CI when runing against the Docker stack:

```
2025-03-25T17:59:07.2742600Z <--- Last few GCs --->
2025-03-25T17:59:07.2742980Z 
2025-03-25T17:59:07.2744315Z [8656:0xf8e4000]  3266845 ms: Scavenge 4045.1 (4127.9) -> 4029.4 (4128.2) MB, pooled: 4 MB, 6.45 / 0.00 ms  (average mu = 0.197, current mu = 0.010) allocation failure; 
2025-03-25T17:59:07.2747185Z [8656:0xf8e4000]  3269306 ms: Mark-Compact 4045.2 (4128.2) -> 4022.5 (4128.2) MB, pooled: 4 MB, 2446.48 / 0.00 ms  (average mu = 0.318, current mu = 0.418) allocation failure; scavenge might not succeed
2025-03-25T17:59:07.2748431Z 
2025-03-25T17:59:07.2748442Z 
2025-03-25T17:59:07.2748678Z <--- JS stacktrace --->
2025-03-25T17:59:07.2749027Z 
2025-03-25T17:59:07.2749973Z FATAL ERROR: Ineffective mark-compacts near heap limit Allocation failed - JavaScript heap out of memory
2025-03-25T17:59:07.2751191Z ----- Native stack trace -----
2025-03-25T17:59:07.2752001Z 
2025-03-25T17:59:07.2756969Z 
2025-03-25T17:59:07.2757329Z <--- Last few GCs --->
2025-03-25T17:59:07.2757849Z 
2025-03-25T17:59:07.2758948Z [8656:0xf8e4000]  3266845 ms: Scavenge 4045.1 (4127.9) -> 4029.4 (4128.2) MB, pooled: 4 MB, 6.45 / 0.00 ms  (average mu = 0.197, current mu = 0.010) allocation failure; 
2025-03-25T17:59:07.2760825Z [8656:0xf8e4000]  3269306 ms: Mark-Compact 4045.2 (4128.2) -> 4022.5 (4128.2) MB, pooled: 4 MB, 2446.48 / 0.00 ms  (average mu = 0.318, current mu = 0.418) allocation failure; scavenge might not succeed
2025-03-25T17:59:07.2762115Z 
2025-03-25T17:59:07.2762125Z 
2025-03-25T17:59:07.2762327Z <--- JS stacktrace --->
2025-03-25T17:59:07.2762565Z 
2025-03-25T17:59:07.2763282Z FATAL ERROR: Ineffective mark-compacts near heap limit Allocation failed - JavaScript heap out of memory
2025-03-25T17:59:07.2764161Z ----- Native stack trace -----
```

Add optimisations to Docker Compose, NodeJs, and Playwright to ensure the most efficient use of resource.

## Changes

- Add the following optimisations to NodeJs in the Dockerfile:
`--optimize-for-size`
- Add Docker Compose optimisation in the package.json when starting docker compose
`COMPOSE_BAKE=true`
- Add optimisations to playwright config:
`args: ['--disable-dev-shm-usage', '--no-sandbox']`

## Validation

Local e2e tests passing in docker
Tested several times in CI and docker e2e tests are passing without memory exceptions